requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
jinja2==3.1.4
python-multipart==0.0.9
requests==2.32.3
pandas==2.2.3
openpyxl==3.1.5
beautifulsoup4==4.12.3
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
python-dotenv==1.0.1
cryptography==43.0.1